GUI Agent/requirements.txt
mss
Pillow
paddleocr
PyYAML
